+++ PrebidMobileTests/RenderingTests/Tests/Prebid/RewardedAdUnitTest.swift
@@ -1,0 +1,36 @@
+/*   Copyright 2018-2024 Prebid.org, Inc.
+ 
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+ 
+  http://www.apache.org/licenses/LICENSE-2.0
+ 
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+  */
+
+import XCTest
+@testable import PrebidMobile
+
+class RewardedAdUnitTest: XCTestCase {
+    
+    func testSetAdPosition() {
+        let adUnit = RewardedAdUnit(configID: "test")
+        
+        let adUnitConfig = adUnit.adUnitConfig
+        
+        adUnit.adPosition = .header
+        
+        XCTAssertEqual(adUnit.adPosition, adUnitConfig.adPosition)
+        XCTAssertEqual(adUnitConfig.adPosition, .header)
+        
+        adUnit.adPosition = .footer
+        
+        XCTAssertEqual(adUnit.adPosition, adUnitConfig.adPosition)
+        XCTAssertEqual(adUnitConfig.adPosition, .footer)
+    }
+}
